refactor(ReservationIntegrationTest): 테스트 분리
