+++ src/test/java/roomescape/integration/AdminIntegrationTest.java
@@ -1,29 +1,14 @@
 package roomescape.integration;
 
-import static org.hamcrest.Matchers.is;
-
 import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import roomescape.storage.ReservationStorage;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class AdminIntegrationTest {
-    @Autowired
-    private ReservationStorage reservationStorage;
-
-    private final List<Long> newReservationsId = new ArrayList<>();
-
     @LocalServerPort
     private int port;
 
@@ -39,76 +24,5 @@ public class AdminIntegrationTest {
                 .when().get("/admin")
                 .then().log().all()
                 .statusCode(200);
-    }
-
-    @Test
-    @DisplayName("관리자 예약 페이지가 잘 접속된다.")
-    void adminReservationPageLoad() {
-        RestAssured.given().log().all()
-                .when().get("/admin/reservation")
-                .then().log().all()
-                .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2023-08-05");
-        params.put("time", "15:40");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("id", is(1));
-
-        newReservationsId.add(1L);
-    }
-
-    @Test
-    @DisplayName("관리자 예약 페이지가 잘 동작한다.")
-    void adminReservationPageWork() {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2023-08-05");
-        params.put("time", "15:40");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("id", is(2));
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-
-        RestAssured.given().log().all()
-                .when().delete("/reservations/2")
-                .then().log().all()
-                .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-    }
-
-    @AfterEach
-    void reset() {
-        for (Long id : newReservationsId) {
-            reservationStorage.delete(id);
-        }
     }
 }
